+++ common/default.nix
@@ -6,7 +6,7 @@
       example = "(import /path/to/obeliskapp {}).exe";
     };
     configSource = lib.mkOption {
-      type = lib.types.path;
+      type = lib.types.str;
       example = "/path/to/obelisk-app/config";
       description = ''
         The path to the obelisk application configuration folder. Configs will be copied from this folder every time the configuration is rebuilt. See <link xlink:heref="https://github.com/obsidiansystems/obelisk/tree/master/skeleton/config#config"> for more information.