+++ user/default.nix
@@ -11,22 +11,13 @@ in
     example = lib.literalExample ''
       obelisks."myapp" = {
         obelisk = (import /path/to/myapp {}).exe;
-        configSource = /path/to/config;
+        configSource = "/path/to/config";
         port = 8080;
       };
     '';
     description = "An attrset of obelisk applications to run as user systemd services.";
   };
-  # Create folders in $XDG_CONFIG_HOME for the obelisks. This is where config
-  # files, symlinks to executables, and symlinks to static assets, will be stored.
-  config.xdg.configFile =
-    lib.mapAttrs'
-      (k: v: {
-        name = "${k}/config";
-        value = {
-          source = v.configSource;
-        };
-      }) obs;
+
   # Configure user systemd services for each obelisk
   config.systemd.user.services = lib.mapAttrs (k: v: {
     Unit.Description = "${k} obelisk application";
@@ -35,12 +26,17 @@ in
     Service.ExecStart = "%S/${k}/backend --port=${builtins.toString v.port}";
     Service.WorkingDirectory = "%S/${k}";
   }) obs;
-  # "Deploy" the application executables
+
+
+  # "Deploy" the application executables and create folders in $XDG_CONFIG_HOME
+  # for the obelisks. This is where config files, symlinks to executables, and
+  # symlinks to static assets, will be stored.
   config.home.activation = lib.mapAttrs'
     (k: v: {
       name = "obelisk-activation-${k}";
       value = config.lib.dag.entryAfter ["writeBoundary"] ''
         ln -sfr ${v.obelisk}/* $XDG_CONFIG_HOME/${k}/
+        ln -sfr ${v.configSource} $XDG_CONFIG_HOME/${k}/config
       '';
     }) obs;
 }