+++ ChangeLog.md
@@ -1,5 +1,11 @@
 # Release history of obelisk-systemd
 
+## v0.2.0.0
+
+* Keep configs out of the nix store
+* system: Make username and home directory configurable
+* system: default to nginx and https off
+
 ## v0.1.0.0
 
 * Provide a user-level systemd service via home-manager.